fix exchange/activesync heartbeat not incrementing properly
The AOSP code only potentially increases the activesync "heartbeat" when the
response from the previous ping was successful AND there were no changes.

This submission allows the heartbeat to also increase if there were changes
AND the full heartbeat time elapsed.  This is in accordance with MS's spec
and allows the device to remain in deep sleep for longer periods of time
when sync'ing with active exchange accounts.  (Examples of this would be a
user deleting items from a sync'd folder via outlook, new contacts, updated
appointments, and any other change other than a new "inbox" email message.)

In addition, this change increases the max heartbeat from google's
completely arbitrary 17 minutes to microsoft's suggested 30 minutes.

Change-Id: I3ea5344d55f00eaf9e0ae270ada9d32c2cec8885
Signed-off-by: Gary Dezern <garyd9@gmail.com>